feat(app): 固定刻みの番号を時間進行配線が所有し、刻み入力へ載せて台本が読む形にする(Issue #50、判断13)
固定刻みの番号(64ビット・飽和なし・剰余を実装しない)を足し、刻みを回す工程が1本ごとに進める。遊ばない起動でも進む。
キツネの場所巡りの配線のゲーム更新の回数(u32の飽和加算)を消し、台本は刻み入力の番号から操作を決める。台本の区切りの
定数は番号の型で持ち、値は変えない。終了時の報告の「ゲーム更新の回数」は起動から進めた刻みの本数を番号から数える。
試験: 番号の型、時間進行配線の番号の進み、台本の各段階が当たる番号、遊ばない起動でも番号が進むこと、
台本の決定が最初の刻みを0と数えた30番目の刻みで効くこと。

+++ crates/blitz_app/src/world_execution/summary.rs
@@ -5,7 +5,7 @@ use super::movement_record::移動の観測の要約;
 /// そのプロセスがゲームをどこまで回したかの要約。
 pub(crate) struct ゲーム進行の要約 {
     pub(crate) ゲームの表示名: &'static str,
-    pub(crate) ゲーム更新の回数: u32, // 1刻みは描画フレーム1回ぶんであるため、ゲームを回した描画フレーム数と一致する
+    pub(crate) ゲーム更新の回数: u64, // 起動から進めた固定刻みの本数。このゲームは起動から全ての刻みで走るため、ゲームを回した刻みの本数と一致する
     pub(crate) 最後の進行段階の呼び名: &'static str,
     pub(crate) 到達済みの目的地数: usize,
     pub(crate) 目的地の総数: usize,
